Moved ExactDepth testing from TestShortestPath to TestExactDepthPathFinder. GraphAlgoFactory now uses ExactDepthPathFinder instead of ShortestPath. ExactDepthPathFinder now works. Performance testings needs to be done.
