+++ community/graph-algo/src/main/java/org/neo4j/graphalgo/impl/path/ExactDepthPathFinder.java
@@ -35,6 +35,7 @@ import org.neo4j.graphdb.traversal.BranchSelector;
 import org.neo4j.graphdb.traversal.TraversalBranch;
 import org.neo4j.graphdb.traversal.TraversalDescription;
 import org.neo4j.graphdb.traversal.Traverser;
+import org.neo4j.graphdb.traversal.Uniqueness;
 
 /**
  * Tries to find paths in a graph from a start node to an end node where the
@@ -53,24 +54,26 @@ public class ExactDepthPathFinder extends TraversalPathFinder
     private final PathExpander expander;
     private final int onDepth;
     private final int startThreshold;
+    private final Uniqueness uniqueness;
 
     public ExactDepthPathFinder( RelationshipExpander expander, int onDepth,
-            int startThreshold )
+            int startThreshold, boolean allowLoops )
     {
-        this( toPathExpander( expander ), onDepth, startThreshold );
+        this( toPathExpander( expander ), onDepth, startThreshold, allowLoops );
     }
 
-    public ExactDepthPathFinder( PathExpander expander, int onDepth, int startThreshold )
+    public ExactDepthPathFinder( PathExpander expander, int onDepth, int startThreshold, boolean allowLoops )
     {
         this.expander = expander;
         this.onDepth = onDepth;
         this.startThreshold = startThreshold;
+        this.uniqueness = allowLoops ? Uniqueness.RELATIONSHIP_GLOBAL : Uniqueness.NODE_PATH;
     }
 
     @Override
     protected Traverser instantiateTraverser( Node start, Node end )
     {
-        TraversalDescription side = traversal().breadthFirst().expand( expander ).uniqueness( RELATIONSHIP_GLOBAL ).order(
+        TraversalDescription side = traversal().breadthFirst().uniqueness( uniqueness ).order(
                 new BranchOrderingPolicy()
                 {
                     public BranchSelector create( TraversalBranch startSource, PathExpander expander )
@@ -81,11 +84,11 @@ public class ExactDepthPathFinder extends TraversalPathFinder
                 } );
         
         return bidirectionalTraversal()
-                .startSide( side.evaluator( toDepth( onDepth/2 ) ) )
-                .endSide( side.evaluator( toDepth( onDepth-onDepth/2 ) ) )
+                .startSide( side.expand( expander ).evaluator( toDepth( onDepth/2 ) ) )
+                .endSide( side.expand( expander.reverse() ).evaluator( toDepth( onDepth-onDepth/2 ) ) )
                 .collisionEvaluator( atDepth( onDepth ) )
                 // TODO Level side selector will make the traversal return wrong result, why? 
 //                .sideSelector( SideSelectorPolicies.LEVEL, onDepth )
                 .traverse( start, end );
-    }
+    }    
 }
